+++ CHANGELOG.md
@@ -5,7 +5,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-- Added method ```root()``` to set root path for templates directory.
+### Added
+- Method ```root()``` to set root path for template directory.
+
+### Fixed
+- Bug when assembling route with optional part and parameter with constraint.
 
 ## [1.0.0] - 2021-02-17
 ### Added